requirements.txt
pip==22.3.1
ez-setup==0.9
setuptools==65.5.1
numpy==1.23.4
spidev==3.6
pytube==12.1.0
wheel==0.38.2
imageio==2.22.3
imageio-ffmpeg==0.4.7
Pillow==9.3.0
RPi.GPIO==0.7.1

